meiduo_py_vue/meiduo/requirements.txt
ipython==8.6.0

async-timeout==4.0.2
Deprecated==1.2.13

django-cors-headers==3.10.1

djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
importlib-metadata==4.8.3

MarkupSafe==2.1.1
packaging==21.3
Pillow==8.4.0

PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2022.6
redis==4.3.4
setuptools==59.6.0
sqlparse==0.4.3


